+++ tests/lib/Unit/ContainerTest.php
@@ -24,7 +24,6 @@ use CloudCreativity\LaravelJsonApi\Contracts\Resolver\ResolverInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Validation\ValidatorFactoryInterface;
 use CloudCreativity\LaravelJsonApi\Exceptions\RuntimeException;
 use Illuminate\Container\Container as IlluminateContainer;
-use League\Flysystem\AdapterInterface;
 use Neomerx\JsonApi\Contracts\Schema\SchemaProviderInterface;
 use PHPUnit\Framework\MockObject\MockObject;
 
@@ -253,7 +252,7 @@ class ContainerTest extends TestCase
 
     public function testResourceAuthorizerIsNotResourceAuthorizer()
     {
-        $mock = $this->createMock(AdapterInterface::class);
+        $mock = $this->createMock(\Illuminate\Contracts\Container\Container::class);
         $this->resolver->method('getAuthorizerByResourceType')->willReturn(get_class($mock));
         $this->illuminateContainer->method('make')->willReturn($mock);